src/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
packaging==23.2
Pillow==10.1.0
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
